[refactor] major refactor of routes
