+++ functions/firebase.js
@@ -1,0 +1,3 @@
+import {initializeApp} from "firebase-admin/app";
+const app = initializeApp();
+export default app;
