+++ functions/index.js
@@ -1,136 +1,18 @@
 /* eslint-disable max-len */
 /* eslint-disable no-unused-vars */
-// import {onRequest} from "firebase-functions/v2/https";
-import {initializeApp} from "firebase-admin/app";
-import {onTaskDispatched} from "firebase-functions/v2/tasks";
-import {getFunctions} from "firebase-admin/functions";
-
-// Possible way to deconstruct index.js into muliple files while maintaining
-// app context:
-// https://github.com/jspsych/datapipe/blob/6240d63c423850e2c5bd2ae691e3a2f3457b171c/functions/index.js
-const app = initializeApp();
-import {onRequest, onCall} from "firebase-functions/v2/https";
-// import {onObjectFinalized} from "firebase-functions/v2/storage";
-import {getAuth} from "firebase-admin/auth";
+import {onCall} from "firebase-functions/v2/https";
 import logger from "firebase-functions/logger";
-import {newUser, validateOnCallAuth, validateOnRequestAdmin} from "./auth/auth.js";
-import {
-  preProcess,
-  hookFromBucket,
-} from "./util/pipeline.js";
+import {validateOnCallAuth} from "./auth/auth.js";
+import {ENVIRONMENT} from "./config/config.js";
 
-import {
-  getUser,
-  getPipelineFirestore,
-  getAiFirestore,
-} from "./storage/firestore.js";
-
-import {
-  libraryAddItemFirestore,
-  libraryGetAllFirestore,
-  libraryDeleteItemFirestore,
-} from "./storage/firestore/library.js";
-
-import {
-  getGlobalScenesFirestore,
-  getCatalogueScenesFirestore,
-  scenesCreateLibraryItemFirestore,
-  scenesUpdateLibraryItemFirestore,
-} from "./storage/firestore/scenes.js";
-
-import {
-  catalogueAddFirestore,
-  catalogueGetAllFirestore,
-  catalogueDeleteFirestore,
-  catalogueUpdateFirestore,
-} from "./storage/firestore/catalogue.js";
-
-import {
-  getAAXAvailableFirestore,
-  setAAXAvailableFirestore,
-  getAAXConnectStatusFirestore,
-} from "./storage/firestore/users.js";
-
-import {
-  generateImages,
-  imageGenRecursive,
-} from "./ai/openai/dallE.js";
-
-import {
-  beforeUserCreated,
-  // beforeUserSignedIn,
-} from "firebase-functions/v2/identity";
-
-import {
-  generatePublicOPDS,
-  generatePrivateOPDS,
-  processRawPublicItem,
-  generateManifest,
-  generateUserItemManifest,
-} from "./util/opds.js";
-
-import {
-  getAAXLoginURL,
-  getAAXAuth,
-  audiblePostAuthHook,
-  refreshAAXTokens,
-  submitAAXAuth,
-  disconnectAAXAuth,
-} from "./util/audibleOpdsHelper.js";
-
-import {
-  ENVIRONMENT,
-  OPENAI_API_KEY,
-  HOSTING_DOMAIN,
-} from "./config/config.js";
-
-import {
-  generateTranscriptions,
-} from "./ai/transcribe.js";
-
-import {
-  getFunctionUrl,
-  dataToBody,
-  largeDispatchInstance,
-  microDispatchInstance,
-  dispatchTask,
-} from "./util/dispatch.js";
-
-import {
-  graphCharacters,
-  graphLocations,
-  graphCharacterDescriptions,
-  graphLocationDescriptions,
-  graphSummarizeDescriptions,
-  graphScenes,
-} from "./ai/graph.js";
-
-import {
-  outpaintWideAndTall,
-} from "./ai/stability/stability.js";
-
-/**
- * Cloud Function triggered before a new user is created.
- * This function logs the event details and can be extended to handle more complex logic.
- *
- * @param {object} event - The event payload containing details about the user being created.
- * @returns {Promise<void>} A promise that resolves when the function is complete.
- *
- * Sample event object:
- *
- */
-
-export const newUserTriggers =
-  beforeUserCreated({region: "europe-west1"}, async (event) => {
-    logger.debug(`FUNCTION: beforeUserCreated - newUserTriggers`);
-    logger.debug(event);
-    try {
-      await newUser(app, event);
-    } catch (error) {
-      logger.error(error);
-    }
-    return;
-  });
+export * from "./routes/users.js";
+export * from "./routes/library.js";
+export * from "./routes/scenes.js";
+export * from "./routes/catalogue.js";
+export * from "./routes/opds.js";
+export * from "./routes/aax.js";
+export * from "./routes/graph.js";
+export * from "./routes/pipeline.js";
 
 /**
  * HTTP Cloud Function test.
@@ -152,309 +34,3 @@ export const helloWorld = onCall({region: "europe-west1"}, async (context) => {
   }
   return {uid: uid, message: `Success! You made an authenticated request.`};
 });
-
-/**
- * Cloud function to get the current auth'd user
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to an object containing the user's UID and the data provided.
- */
-export const getCurrentUser = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return getUser(uid);
-});
-
-/**
- * Cloud Function to create a new book entry.
- * This function is triggered by an on-call request and requires the user to be authenticated.
- *
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to an object containing the user's UID and the data provided.
- */
-export const v1addItemToLibrary = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return libraryAddItemFirestore(uid, data, app);
-});
-
-/**
- * Cloud Function to get the manifest for a specific item in the user's library.
- * This function is triggered by an on-call request and requires the user to be authenticated.
- *
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to the item manifest if found and the user is authenticated.
- */
-export const v1getItemManifest = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return generateUserItemManifest(app, uid, data);
-});
-
-/**
- * Retrieves a book from the Firestore database based on the user's UID and the book ID provided in the data.
- * This function is triggered by an on-call request and requires the user to be authenticated.
- *
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to the book data if found and the user is authenticated, otherwise null.
- */
-export const v1getLibrary = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return libraryGetAllFirestore(app, uid, data);
-});
-
-/**
- * Requests the server delete a book, including any items in storage
- * This function is triggered by an on-call request and requires the user to be authenticated.
- *
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to the book data if found and the user is authenticated, otherwise null.
- */
-export const v1deleteItemsFromLibrary = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return libraryDeleteItemFirestore(uid, data, app);
-});
-
-export const getPipeline = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return getPipelineFirestore(uid, data, app);
-});
-
-/**
- * Retrieves a pipeline from the Firestore database based on the user's UID and the pipeline ID provided in the data.
- * This function is triggered by an on-call request and requires the user to be authenticated.
- *
- * @param {object} context - The context object provided by Firebase Functions, containing authentication details and data.
- * @returns {Promise<object>} A promise that resolves to the pipeline data if found and the user is authenticated, otherwise null.
- */
-export const preProcessBook = onRequest({region: "europe-west1"},
-    // {cors: [/firebase\.com$/, "flutter.com"]},
-    async (req, res) => {
-      return preProcess(req, res);
-    });
-
-export const v1catalogueAdd = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await catalogueAddFirestore(req, app));
-});
-
-export const v1catalogueGet = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return catalogueGetAllFirestore(app);
-});
-
-export const v1catalogueGetOPDS = onRequest({region: "europe-west1"}, async (req, res) => {
-  res.status(200).send(await generatePublicOPDS(app));
-});
-
-export const v1catalogueGetManifest = onRequest({region: "europe-west1"}, async (req, res) => {
-  const catalogueId = req.path.split("/").pop();
-  if (!catalogueId) {
-    res.status(400).send("Catalogue ID is required");
-    return;
-  }
-  console.log(`catalogueId: ${catalogueId}`);
-  res.status(200).send(await generateManifest(app, "admin", catalogueId));
-});
-
-export const v1catalogueDelete = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await catalogueDeleteFirestore(req, app));
-});
-
-export const v1catalogueUpdate = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await catalogueUpdateFirestore(req, app));
-});
-
-// /v1/ai/generateImages
-export const v1generateSceneImages = onRequest({region: "europe-west1", cors: true}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  await dispatchTask("generateSceneImages", req.body, 60 * 5, 1);
-  res.status(200).send({dispatched: true});
-});
-
-export const v1getAi = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getAiFirestore(uid, data, app);
-});
-
-export const v1getLibraryScenes = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getGlobalScenesFirestore(uid, data, app);
-});
-
-export const v1getCatalogueItemScenes = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getCatalogueScenesFirestore(uid, data, app);
-});
-
-
-export const v1addLibraryItemScenes = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await scenesCreateLibraryItemFirestore(uid, data, app);
-});
-
-export const v1updateLibraryItemScenes = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await scenesUpdateLibraryItemFirestore(uid, data, app);
-});
-
-// Endpoints to use audible-opds-firebase
-export const v1getAAXLoginURL = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getAAXLoginURL(uid, data, app);
-});
-
-export const v1aaxConnect = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  const auth = await getAAXAuth(uid, data, app);
-  await dispatchTask("aaxPostAuthHook", {uid: uid, auth: auth});
-  return auth;
-});
-
-export const v1AdminSubmitAAXAuth = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await submitAAXAuth(req, app));
-});
-
-// export const v1TMPaudiblePostAuthHook = onCall({
-//   region: "europe-west1",
-//   memory: "32GiB",
-//   concurrency: 1,
-//   timeoutSeconds: 540,
-// }, async (context) => {
-//   const {uid, data} = await validateOnCallAuth(context);
-//   return await audiblePostAuthHook(uid, data, app);
-// });
-
-export const v1getAAXConnectStatus = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getAAXConnectStatusFirestore(uid);
-});
-
-export const v1disconnectAAX = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await disconnectAAXAuth(uid, data, app);
-});
-
-export const v1refreshAAXTokens = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await refreshAAXTokens(data);
-});
-
-export const v1generateTranscriptions = onCall({
-  region: "europe-west1",
-  memory: "32GiB",
-  concurrency: 1,
-  timeoutSeconds: 540,
-}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await generateTranscriptions(uid, data, app);
-});
-
-export const v1catalogueProcessRaw = onRequest({
-  region: "europe-west1",
-}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await dispatchTask("processM4B", {sku: req.body.sku}));
-});
-
-export const v1getAAXAvailable = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await getAAXAvailableFirestore(uid, data, app);
-});
-
-export const v1AdminSetAAXAvailable = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await setAAXAvailableFirestore(req, app));
-});
-
-export const v1getPrivateOPDSFeed = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return await generatePrivateOPDS(uid, data, app);
-});
-
-export const v1getPrivateOPDSFeedURL = onCall({region: "europe-west1"}, async (context) => {
-  const {uid, data} = await validateOnCallAuth(context);
-  return {url: `${HOSTING_DOMAIN.value()}/v1/tmp/privateOPDS/${uid}`};
-});
-
-export const v1TMPgetPrivateOPDSFeed = onRequest({region: "europe-west1"}, async (req, res) => {
-  const uid = req.path.split("/").pop();
-  res.status(200).send(await generatePrivateOPDS(uid, req.body, app));
-});
-
-export const v1TMPgetPrivateManifest = onRequest({region: "europe-west1"}, async (req, res) => {
-  const pathParts = req.path.split("/");
-  const uid = pathParts[pathParts.length - 2];
-  const catalogueId = pathParts[pathParts.length - 1];
-  res.status(200).send(await generateManifest(app, uid, catalogueId));
-});
-
-export const v1AdminGraphCharacters = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await graphCharacters(app, req));
-});
-
-export const v1AdminGraphLocations = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await graphLocations(app, req));
-});
-
-export const v1AdminOutpaintImage = onRequest({region: "europe-west1"}, async (req, res) => {
-  await validateOnRequestAdmin(req);
-  res.status(200).send(await outpaintWideAndTall(app, req.body));
-});
-
-// Dispatch Tasks.
-
-export const generateGraphCharacterDescriptions = onTaskDispatched(
-    microDispatchInstance(),
-    async (req) => {
-      logger.debug(`graphCharacterDescriptions: ${JSON.stringify(req.data)}`);
-      return await graphCharacterDescriptions(app, dataToBody(req));
-    });
-
-export const generateGraphLocationDescriptions = onTaskDispatched(
-    microDispatchInstance(),
-    async (req) => {
-      logger.debug(`graphLocationDescriptions: ${JSON.stringify(req.data)}`);
-      return await graphLocationDescriptions(app, dataToBody(req));
-    });
-
-export const generateGraphSummarizeDescriptions = onTaskDispatched(
-    microDispatchInstance(),
-    async (req) => {
-      logger.debug(`graphSummarizeDescriptions: ${JSON.stringify(req.data)}`);
-      return await graphSummarizeDescriptions(app, dataToBody(req));
-    });
-
-export const generateGraphScenes = onTaskDispatched(
-    microDispatchInstance(),
-    async (req) => {
-      logger.debug(`graphScenes: ${JSON.stringify(req.data)}`);
-      return await graphScenes(app, dataToBody(req));
-    });
-
-export const aaxPostAuthHook = onTaskDispatched(
-    largeDispatchInstance(),
-    async (req) => {
-      // logger.debug(`aaxPostAuthHook: ${JSON.stringify(req.data)}`);
-      // const body = req.data;
-      const {body: body} = dataToBody(req);
-      return await audiblePostAuthHook(body.uid, {auth: body.auth}, app);
-    },
-);
-
-export const processM4B = onTaskDispatched(
-    largeDispatchInstance(),
-    async (req) => {
-      logger.debug(`processM4B: ${JSON.stringify(req.data)}`);
-      return await processRawPublicItem(dataToBody(req), app);
-    },
-);
-
-export const generateSceneImages = onTaskDispatched(
-    largeDispatchInstance(),
-    async (req) => {
-      logger.debug(`generateSceneImages: ${JSON.stringify(req.data)}`);
-      return await imageGenRecursive(dataToBody(req), app);
-    },
-);
